+++ Source/EventFlow.Tests/IntegrationTests/InMemoryConfiguration.cs
@@ -34,34 +34,34 @@ namespace EventFlow.Tests.IntegrationTests
 {
     public class InMemoryConfiguration : IntegrationTestConfiguration
     {
-        private IInMemoryReadModelStore<TestAggregateReadModel> _inMemoryReadModelStore;
+        private IInMemoryReadModelStore<InMemoryTestAggregateReadModel> _inMemoryReadModelStore;
         private IReadModelPopulator _readModelPopulator;
 
         public override IRootResolver CreateRootResolver(EventFlowOptions eventFlowOptions)
         {
             var resolver = eventFlowOptions
-                .UseInMemoryReadStoreFor<TestAggregateReadModel, ILocateByAggregateId>()
+                .UseInMemoryReadStoreFor<InMemoryTestAggregateReadModel, ILocateByAggregateId>()
                 .CreateResolver();
 
-            _inMemoryReadModelStore = resolver.Resolve<IInMemoryReadModelStore<TestAggregateReadModel>>();
+            _inMemoryReadModelStore = resolver.Resolve<IInMemoryReadModelStore<InMemoryTestAggregateReadModel>>();
             _readModelPopulator = resolver.Resolve<IReadModelPopulator>();
 
             return resolver;
         }
 
-        public override Task<ITestAggregateReadModel> GetTestAggregateReadModelAsync(IIdentity id)
+        public override async Task<ITestAggregateReadModel> GetTestAggregateReadModelAsync(IIdentity id)
         {
-            return Task.FromResult<ITestAggregateReadModel>(_inMemoryReadModelStore.Get(id));
+            return await _inMemoryReadModelStore.GetByIdAsync(id.Value, CancellationToken.None).ConfigureAwait(false);
         }
 
         public override Task PurgeTestAggregateReadModelAsync()
         {
-            return _readModelPopulator.PurgeAsync<TestAggregateReadModel>(CancellationToken.None);
+            return _readModelPopulator.PurgeAsync<InMemoryTestAggregateReadModel>(CancellationToken.None);
         }
 
         public override Task PopulateTestAggregateReadModelAsync()
         {
-            return _readModelPopulator.PopulateAsync<TestAggregateReadModel>(CancellationToken.None);
+            return _readModelPopulator.PopulateAsync<InMemoryTestAggregateReadModel>(CancellationToken.None);
         }
 
         public override void TearDown()