+++ Source/EventFlow/EventStores/InMemory/InMemoryEventStore.cs
@@ -143,6 +143,23 @@ namespace EventFlow.EventStores.InMemory
             }
         }
 
+        public override Task DeleteAggregateAsync<TAggregate, TIdentity>(
+            TIdentity id,
+            CancellationToken cancellationToken)
+        {
+            if (_eventStore.ContainsKey(id.Value))
+            {
+                List<ICommittedDomainEvent> committedDomainEvents;
+                _eventStore.TryRemove(id.Value, out committedDomainEvents);
+                Log.Verbose(
+                    "Deleted aggregate '{0}' with ID '{1}' by deleting all of its {2} events",
+                    typeof(TAggregate).Name,
+                    id,
+                    committedDomainEvents.Count);
+            }
+
+            return Task.FromResult(0);
+        }
         protected override Task<IReadOnlyCollection<ICommittedDomainEvent>> LoadCommittedEventsAsync(
             GlobalSequenceNumberRange globalSequenceNumberRange,
             CancellationToken cancellationToken)