+++ Source/EventFlow/ReadStores/InMemory/InMemoryReadModelStore.cs
@@ -27,6 +27,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using EventFlow.Aggregates;
 using EventFlow.Logs;
+using EventFlow.Queries;
 
 namespace EventFlow.ReadStores.InMemory
 {
@@ -79,10 +80,17 @@ namespace EventFlow.ReadStores.InMemory
             return _readModels.Values;
         }
 
-        public IEnumerable<TReadModel> Find(Func<TReadModel, bool> predicate)
+        public IEnumerable<TReadModel> Find(Predicate<TReadModel> predicate)
         {
-            return _readModels.Values
-                .Where(predicate);
+            return _readModels.Values.Where(rm => predicate(rm));
+        }
+		
+		public override Task<TReadModel> GetByIdAsync(string id, CancellationToken cancellationToken)
+        {
+            TReadModel readModel;
+            return _readModels.TryGetValue(id, out readModel)
+                ? Task.FromResult(readModel)
+                : Task.FromResult(default(TReadModel));
         }
 
         public override Task PurgeAsync<TReadModelToPurge>(CancellationToken cancellationToken)