+++ Source/EventFlow/ReadStores/InMemory/Queries/InMemoryQuery.cs
@@ -20,29 +20,20 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-using EventFlow.Configuration.Registrations;
+using System;
+using System.Collections.Generic;
 using EventFlow.Queries;
-using EventFlow.ReadStores.MsSql.Queries;
 
-namespace EventFlow.ReadStores.MsSql.Extensions
+namespace EventFlow.ReadStores.InMemory.Queries
 {
-    public static class EventFlowOptionsExtensions
+    public class InMemoryQuery<TReadModel> : IQuery<IEnumerable<TReadModel>>
+        where TReadModel : IReadModel, new()
     {
-        public static EventFlowOptions UseMssqlReadModel<TReadModel, TReadModelLocator>(this EventFlowOptions eventFlowOptions)
-            where TReadModel : IMssqlReadModel, new()
-            where TReadModelLocator : IReadModelLocator
-        {
-            eventFlowOptions.RegisterServices(f =>
-                {
-                    if (!f.HasRegistrationFor<IReadModelSqlGenerator>())
-                    {
-                        f.Register<IReadModelSqlGenerator, ReadModelSqlGenerator>(Lifetime.Singleton);
-                    }
-                    f.Register<IReadModelStore, MssqlReadModelStore<TReadModel, TReadModelLocator>>();
-                    f.Register<IQueryHandler<ReadModelByIdQuery<TReadModel>, TReadModel>, MsSqlReadModelByIdQueryHandler<TReadModel>>();
-                });
+        public Predicate<TReadModel> Query { get; private set; }
 
-            return eventFlowOptions;
+        public InMemoryQuery(Predicate<TReadModel> query)
+        {
+            Query = query;
         }
     }
 }